+++ ice-20201109/alibabacloud_ice20201109/models/_clear_aiagent_voiceprint_request.py
@@ -7,8 +7,10 @@ from darabonba.model import DaraModel
 class ClearAIAgentVoiceprintRequest(DaraModel):
     def __init__(
         self,
+        registration_mode: str = None,
         voiceprint_id: str = None,
     ):
+        self.registration_mode = registration_mode
         # The unique identifier for the voiceprint.
         self.voiceprint_id = voiceprint_id
 
@@ -20,6 +22,9 @@ class ClearAIAgentVoiceprintRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.registration_mode is not None:
+            result['RegistrationMode'] = self.registration_mode
+
         if self.voiceprint_id is not None:
             result['VoiceprintId'] = self.voiceprint_id
 
@@ -27,6 +32,9 @@ class ClearAIAgentVoiceprintRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('RegistrationMode') is not None:
+            self.registration_mode = m.get('RegistrationMode')
+
         if m.get('VoiceprintId') is not None:
             self.voiceprint_id = m.get('VoiceprintId')