+++ ice-20201109/alibabacloud_ice20201109/models/_list_aiagent_voiceprints_request.py
@@ -9,12 +9,14 @@ class ListAIAgentVoiceprintsRequest(DaraModel):
         self,
         page_number: int = None,
         page_size: int = None,
+        registration_mode: str = None,
         voiceprint_id: str = None,
     ):
         # The page number.
         self.page_number = page_number
         # The number of entries per page. Value values: [1,100].
         self.page_size = page_size
+        self.registration_mode = registration_mode
         # A unique identifier for the voiceprint. This parameter is optional. If provided, only the information for that ID is returned. If not specified, all voiceprints under the account are returned.
         self.voiceprint_id = voiceprint_id
 
@@ -32,6 +34,9 @@ class ListAIAgentVoiceprintsRequest(DaraModel):
         if self.page_size is not None:
             result['PageSize'] = self.page_size
 
+        if self.registration_mode is not None:
+            result['RegistrationMode'] = self.registration_mode
+
         if self.voiceprint_id is not None:
             result['VoiceprintId'] = self.voiceprint_id
 
@@ -44,6 +49,9 @@ class ListAIAgentVoiceprintsRequest(DaraModel):
 
         if m.get('PageSize') is not None:
             self.page_size = m.get('PageSize')
+
+        if m.get('RegistrationMode') is not None:
+            self.registration_mode = m.get('RegistrationMode')
 
         if m.get('VoiceprintId') is not None:
             self.voiceprint_id = m.get('VoiceprintId')